Extracted AddEditGameForm from EventAddEditGame
